Split explog test from test_math.py
